Fixed makefile and <immintrin> includes
For some reason my compiler DID NOT like when I left it at <immintrin> so it's now <imintrin.h>

The makefile now builds a working library, properly, and is usable in any C++ code! I struggled with the Keccak build. The name space declaration had to be moved from source to header, and some function calls modified to reflect that.