Updated several files, and fixed the RNG issue
I was having issues with keccak again, but those are fixed finally.

The RNG issue is fixed, and now instead of using a pointer to a function, the RNG utilizes a shared pointer to a subclass of the random class. This enables better polymorphism and closer to true C++ functionality. There may be some debugging statements left in the code, will fix in the next commit.